v0fix: fix production path error
Paths are now resolved relative to the directory. Not doing so means the
absolute path gets baked in and if the path is slightly different in the actual
prod server (as compared to the build server) we get issues.

If you pass: ~~/db it will now resolve to db both during dev
and production. The module will resolve them again internally when it needs
them.